add validation message in logout

+++ src/components/custom/navbar/HorizontalMenu.vue
@@ -53,9 +53,23 @@
 export default {
 	methods: {
 		doLogout(routeTo) {
-			if (routeTo === "login.employee") {
-				localStorage.clear();
-			}
+			this.$swal({
+				title: "Warning",
+				text: "Are you sure want to log out?",
+				type: "warning",
+				showCancelButton: true,
+				confirmButtonColor: "#3085d6",
+				cancelButtonColor: "#d33",
+				confirmButtonText: "Yes!",
+				cancelButtonText: "No",
+			}).then((result) => {
+				if (result.value) {
+					if (routeTo === "login.employee") {
+						router.replace({ name: "login.employee" });
+						localStorage.clear();
+					}
+				}
+			});
 		},
 	},
 };
